// src/types/details.ts
interface Genre {
  id: number;
  name: string;
}

interface ProductionCompany {
  id: number;
  logoPath?: string;
  name: string;
  originCountry: string;
}

interface ProductionCountry {
  iso31661: string;
  name: string;
}

interface SpokenLanguage {
  englishName: string;
  iso6391: string;
  name: string;
}

interface Credits {
  cast: {
    adult: boolean;
    gender: number;
    id: number;
    knownForDepartment: string;
    name: string;
    originalName: string;
    popularity: number;
    profilePath?: string;
    character: string;
    creditId: string;
    order: number;
    castId?: number;
  }[];
  crew: {
    adult: boolean;
    gender: number;
    id: number;
    knownForDepartment: string;
    name: string;
    originalName: string;
    popularity: number;
    profilePath?: string;
    creditId: string;
    department: string;
    job: string;
  }[];
}

interface MediaBase {
  adult: boolean;
  backdropPath: string;
  genres: Genre[];
  homepage: string;
  id: number;
  originCountry: string[];
  originalLanguage: string;
  overview: string;
  popularity: number;
  posterPath: string;
  productionCompanies: ProductionCompany[];
  productionCountries: ProductionCountry[];
  releaseDate: string | null;
  spokenLanguages: SpokenLanguage[];
  status: string;
  tagline: string;
  voteAverage: number;
  voteCount: number;
  credits: Credits;
}

export interface TVDetailsType extends MediaBase {
  mediaType: "tv";
  creator: string;
  creatorList: {
    id: number;
    creditId: string;
    name: string;
    originalName: string;
    gender: number;
    profilePath: string | null;
  }[];
  episodeRunTime: number[];
  inProduction: boolean;
  languages: string[];
  lastAirDate: string;
  lastEpisodeToAir: {
    id: number;
    name: string;
    overview: string;
    voteAverage: number;
    voteCount: number;
    airDate: string;
    episodeNumber: number;
    episodeType: string;
    productionCode: string;
    runtime: number;
    seasonNumber: number;
    showId: number;
    stillPath: string;
  };
  nextEpisodeToAir: any;
  networks: {
    id: number;
    logoPath: string;
    name: string;
    originCountry: string;
  }[];
  numberOfEpisodes: number;
  numberOfSeasons: number;
  originalTitle: string;
  releaseDate: string;
  seasons: {
    airDate: string;
    episodeCount: number;
    id: number;
    name: string;
    overview: string;
    posterPath: string;
    seasonNumber: number;
    voteAverage: number;
  }[];
  title: string;
  type: string;
}

export interface MovieDetailsType extends MediaBase {
  mediaType: "movie";
  belongsToCollection: any;
  budget: number;
  director: string;
  directorList: {
    adult: boolean;
    id: number;
    name: string;
    popularity: number;
    gender: number;
    known_for_department: string;
    original_name: string;
    profile_path: string | null;
    credit_id: string;
    department: string;
    job: string;
  }[];
  imdbId: string;
  originalTitle: string;
  revenue: number;
  runtime: number;
  title: string;
  video: boolean;
}